% Auto-generated: do not edit by hand
\name{dashReactJsonSchemaForm}

\alias{dashReactJsonSchemaForm}

\title{DashReactJsonSchemaForm component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashReactJsonSchemaForm(id=NULL, className=NULL, formData=NULL, label=NULL,
schema=NULL, style=NULL, uiSchema=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Sets the class name of the element (the value of an element's html
class attribute).}

\item{formData}{Named list. The formData.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{schema}{Named list. The schema in the input.}

\item{style}{Named list. Add inline styles to the root element.}

\item{uiSchema}{Named list. The uiSchema in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

